+++ src/main/java/kr/hhplus/be/commerce/interfaces/api/BalanceController.java
@@ -6,9 +6,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
 import kr.hhplus.be.commerce.app.BalanceFacade;
-import kr.hhplus.be.commerce.interfaces.dto.requestDto.ChargeBalanceRequest;
-import kr.hhplus.be.commerce.interfaces.dto.requestDto.UserBalanceRequest;
-import kr.hhplus.be.commerce.interfaces.dto.responseDto.ResponseDto;
+import kr.hhplus.be.commerce.interfaces.dto.request.ChargeBalanceRequest;
+import kr.hhplus.be.commerce.interfaces.dto.request.UserBalanceRequest;
+import kr.hhplus.be.commerce.interfaces.dto.response.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
