+++ src/main/java/kr/hhplus/be/commerce/interfaces/api/CouponController.java
@@ -7,8 +7,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
 import kr.hhplus.be.commerce.app.CouponFacade;
 import kr.hhplus.be.commerce.app.dto.CouponResponse;
-import kr.hhplus.be.commerce.interfaces.dto.requestDto.CouponIssueRequest;
-import kr.hhplus.be.commerce.interfaces.dto.responseDto.ResponseDto;
+import kr.hhplus.be.commerce.interfaces.dto.request.CouponIssueRequest;
+import kr.hhplus.be.commerce.interfaces.dto.response.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
