+++ src/main/java/kr/hhplus/be/commerce/interfaces/api/ProductController.java
@@ -8,8 +8,8 @@ import jakarta.validation.Valid;
 import kr.hhplus.be.commerce.app.ProductFacade;
 import kr.hhplus.be.commerce.app.dto.ProductResponse;
 import kr.hhplus.be.commerce.app.dto.TopProductResponse;
-import kr.hhplus.be.commerce.interfaces.dto.requestDto.ProductPageRequest;
-import kr.hhplus.be.commerce.interfaces.dto.responseDto.ResponseDto;
+import kr.hhplus.be.commerce.interfaces.dto.request.ProductPageRequest;
+import kr.hhplus.be.commerce.interfaces.dto.response.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
