+++ src/main/java/kr/hhplus/be/commerce/interfaces/dto/request/UserBalanceRequest.java
@@ -1,5 +1,6 @@
-package kr.hhplus.be.commerce.interfaces.dto.requestDto;
+package kr.hhplus.be.commerce.interfaces.dto.request;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Builder;
@@ -8,6 +9,7 @@ import lombok.Builder;
 public record UserBalanceRequest(
         @NotNull(message = "사용자 ID는 필수값입니다.")
         @Positive(message = "유효하지 않은 사용자 ID입니다.")
+        @Schema(description = "사용자 ID")
         Long userId
 ) {
 }
