--- src/main/java/kr/hhplus/be/commerce/interfaces/dto/requestDto/OrderItemRequest.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.commerce.interfaces.dto.requestDto;
-
-import lombok.Builder;
-
-@Builder
-public record OrderItemRequest(
-        Long productId,
-        Long quantity
-) {
-}
