+++ src/main/java/kr/hhplus/be/commerce/interfaces/dto/response/ResponseDto.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.commerce.interfaces.dto.responseDto;
+package kr.hhplus.be.commerce.interfaces.dto.response;
 
 import lombok.Builder;
